Cleanup function code, add copy button index (#10)
### Summary
After examining all the code, the root cause of the copy button bug was
discovered to be the previous removal of the `sidebar-resize-handle`.
One of the core files mdbook generates, `book.js`, was looking for this
element and failing silently, causing an uncaught error.

The solution is to restore the element and hide it instead to avoid
writing a separate copy function to bypass default behavior.

In addition to this cleanup, this PR also adds indexing by class for the
copy buttons, so that each button can be uniquely identified (for
analytics).